+++ Code/ConfigOptions.cs
@@ -2,20 +2,25 @@
 using System.Collections.Generic;
 using System.Text;
 using BepInEx.Configuration;
+using MiscFixes.Modules;
 
 namespace DamageSourceForEquipment
 {
     public static class ConfigOptions
     {
-        public static ConfigEntry<bool> AspectPassiveDamageIsEquipment;
+        public static ConfigEntry<bool> AspectDamageIsEquipment;
 
         internal static void BindConfigOptions(ConfigFile config)
         {
-            AspectPassiveDamageIsEquipment = config.Bind<bool>(
+            AspectDamageIsEquipment = config.BindOption<bool>(
                 "Aspects",
-                "Aspect passive effect damage", false,
-                "Should damage from aspect's passive effects (i.e blazing fire trail, malachite spikes) count as equipment damage? This is configurable since Wake of Vultures makes the aspect effects not tied to an equipment."
+                "Make aspect effect damage count as equipment damage",
+                "Should damage from aspect's passive effects (i.e blazing fire trail, malachite spikes) count as equipment damage?",
+                true,
+                Extensions.ConfigFlags.RestartRequired
             );
+
+            config.WipeConfig();
         }
     }
 }