+++ Thunderstore/Package Files/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0
+
+- Fixed for AC release
+- Made overloading orbs count as equipment damage
+- "Aspect damage is equipment" config is true by default
+- Ported to MonoDetour
+
 ## 1.0.2
 
 - Fixed WolfoItemBuffs' pocket ICBM change breaking this mod's disposable missile launcher change